refactor(tui): split mod.rs into dm.rs, frame.rs, event_loop.rs (#584)
Extract DM room handling (open_dm_tab, handle_dm_action, switch_to_tab,
write_payload_to_tab, DmTabConfig, read_user_token) to tui/dm.rs.

Extract terminal draw closure body (DrawContext struct, draw_frame
function) to tui/frame.rs.

Extract run() function and stderr redirect helpers to
tui/event_loop.rs.

mod.rs retains RoomTab, DrainResult, process_message/drain_messages
impl, and all tests (~119 prod lines, down from 976). All files under
500 lines. No API changes — run() re-exported at the same path.

Closes #584
